first call working without need for refresh of the page
