+++ src/main.js
@@ -14,10 +14,10 @@ import 'vue-sidebar-menu/dist/vue-sidebar-menu.css';
 
 import { createPinia } from 'pinia';
 
-const pinia = createPinia();
-
 /* import font awesome icon component */
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
+
+import { apolloClient } from './apolloClient';
 
 /* import the fontawesome core */
 import { library } from '@fortawesome/fontawesome-svg-core';
@@ -66,42 +66,7 @@ library.add(
 
 import { DefaultApolloClient } from '@vue/apollo-composable';
 
-import {
-  ApolloClient,
-  createHttpLink,
-  InMemoryCache,
-  ApolloLink,
-  from,
-} from '@apollo/client/core';
-
-const httpLink = createHttpLink({
-  // You should use an absolute URL here
-  uri: 'http://localhost:5000/graphql',
-});
-
-let token = localStorage.getItem('authToken');
-
-const additiveLink = from([
-  new ApolloLink((operation, forward) => {
-    operation.setContext(({ headers }) => ({
-      headers: {
-        ...headers,
-        authtoken: token ? token : null,
-      },
-    }));
-    return forward(operation); // Go to the next link in the chain. Similar to `next` in Express.js middleware.
-  }),
-  httpLink,
-]);
-
-// Cache implementation
-const cache = new InMemoryCache();
-
-// Create the apollo client
-const apolloClient = new ApolloClient({
-  link: additiveLink,
-  cache,
-});
+const pinia = createPinia();
 
 const app = createApp({
   setup() {
